# contracts/http/swagger.yaml
basePath: /
definitions:
...
  model.ActionExecutionStatus:
    enum:
    - pending
    - running
    - succeeded
    - failed
    - retrying
    - skipped
    - cancelled
    type: string
    x-enum-varnames:
    - ActionExecutionPending
    - ActionExecutionRunning
    - ActionExecutionSucceeded
    - ActionExecutionFailed
    - ActionExecutionRetrying
    - ActionExecutionSkipped
    - ActionExecutionCancelled
  model.ActionType:
    enum:
    - send_dm
    - timeout_user
    - kick_user
    - ban_user
    - remove_timeout
    - unban_user
    type: string
    x-enum-varnames:
    - ActionSendDM
    - ActionTimeoutUser
    - ActionKickUser
    - ActionBanUser
    - ActionRemoveTimeout
    - ActionUnbanUser
...
  quack.GuildSettingsInput:
    properties:
      audit_mirror_channel_discord_id:
        type: string
      general_logging_enabled:
        type: boolean
      honeypot_enabled:
        type: boolean
      managed_evidence_channel_discord_id:
        type: string
      notification_footer:
        type: string
      notification_introduction:
        type: string
      tickets_enabled:
        type: boolean
    type: object
...
  quack.OpsActionCapability:
    properties:
      action_type:
        $ref: '#/definitions/model.ActionType'
      executable:
        type: boolean
      status:
        type: string
    type: object
  quack.OpsActionStatus:
    properties:
      capabilities:
        items:
          $ref: '#/definitions/quack.OpsActionCapability'
        type: array
      oldest_pending_or_retry:
        $ref: '#/definitions/quack.OpsOldestActionExecution'
      recent_failures:
        items:
          $ref: '#/definitions/quack.OpsRecentActionFailure'
        type: array
      status_counts:
        additionalProperties:
          format: int64
          type: integer
        type: object
    type: object
  quack.OpsOldestActionExecution:
    properties:
      action_type:
        $ref: '#/definitions/model.ActionType'
      case_id:
        type: string
      case_number:
        type: integer
      created_at:
        type: string
      id:
        type: string
      next_retry_at:
        type: string
      status:
        $ref: '#/definitions/model.ActionExecutionStatus'
    type: object
  quack.OpsRecentActionFailure:
    properties:
      action_type:
        $ref: '#/definitions/model.ActionType'
      case_id:
        type: string
      case_number:
        type: integer
      id:
        type: string
      last_error:
        type: string
      last_error_code:
        type: string
      status:
        $ref: '#/definitions/model.ActionExecutionStatus'
      updated_at:
        type: string
    type: object
  quack.OpsStatusResponse:
    properties:
      actions:
        $ref: '#/definitions/quack.OpsActionStatus'
      generated_at:
        type: string
      guild_id:
        type: string
      queue:
        $ref: '#/definitions/quack.QueueStats'
      scope:
        type: string
    type: object
  quack.QueueStats:
    properties:
      active:
        type: boolean
      buffer_size:
        type: integer
      dropped_total:
        type: integer
      enqueued_total:
        type: integer
      failed_total:
        type: integer
      last_processed_id:
        type: string
      last_processed_type:
        type: string
      panicked_total:
        type: integer
      processed_total:
        type: integer
      queue_size:
        type: integer
      workers:
        type: integer
    type: object
...
info:
  contact: {}
  description: HTTP boundary exposed by the Quack v5 backend.
  title: Quack HTTP API
  version: "5.0"
paths:
...
  /guilds/{discordGuildID}/ops/status:
    get:
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get guild operations status
      tags:
      - Operations
  /guilds/{discordGuildID}/settings:
    get:
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get guild settings
      tags:
      - Guild settings
    patch:
      consumes:
      - application/json
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      - description: Retry-safe request key
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Partial guild settings
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/quack.GuildSettingsInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Update guild settings
      tags:
      - Guild settings
  /guilds/{discordGuildID}/settings/starter-policy-notice/acknowledge:
    post:
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      - description: Retry-safe request key
        in: header
        name: Idempotency-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Acknowledge the starter policy notice
      tags:
      - Guild settings
...
  /ops/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quack.OpsStatusResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - OpsKey: []
      summary: Get global operations status
      tags:
      - Operations
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    in: header
    name: Cookie
    type: apiKey
  MetricsKey:
    in: header
    name: X-Quack-Metrics-Key
    type: apiKey
  OpsKey:
    in: header
    name: X-Quack-Ops-Key
    type: apiKey
swagger: "2.0"